rename filename to base and use BaseEffect
